refactor: 움직임 전략 구현 방법 수정
- 기존 enum으로 관리했던 움직임 전략 부분을 인터페이스를 구현하는 형태로 변경합니다.
- enum으로 관리할 경우 추가되는 전략이 있을 때마다 코드가 길어지고,
다른 매개변수가 필요한 전략이 필요한 경우 불필요한 코드가 늘어날 수 있기에 변경합니다.
