+++ src/main/java/racingcar/model/strategy/AlwaysMoveStrategy.java
@@ -1,0 +1,10 @@
+package racingcar.model.strategy;
+
+public class AlwaysMoveStrategy implements MoveStrategy {
+
+    @Override
+    public boolean canMove() {
+        return true;
+    }
+
+}
