+++ src/main/java/racingcar/model/strategy/MoveStrategy.java
@@ -1,0 +1,5 @@
+package racingcar.model.strategy;
+
+public interface MoveStrategy {
+    boolean canMove();
+}
