+++ src/test/java/racingcar/model/RacingTest.java
@@ -2,6 +2,8 @@ package racingcar.model;
 
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
+import racingcar.model.strategy.AlwaysMoveStrategy;
+import racingcar.model.strategy.MoveStrategy;
 
 import java.util.List;
 
@@ -11,11 +13,13 @@ class RacingTest {
 
     private Car testCar1;
     private Car testCar2;
+    private MoveStrategy moveStrategy;
 
     @BeforeEach
     void setUp() {
         testCar1 = new Car("pobi");
         testCar2 = new Car("woni");
+        moveStrategy = new AlwaysMoveStrategy();
     }
 
     @Test
@@ -24,7 +28,7 @@ class RacingTest {
         Cars cars = new Cars(List.of(testCar1, testCar2));
 
         // when
-        Racing racing = new Racing(cars, MoveStrategy.ALWAYS_MOVE);
+        Racing racing = new Racing(cars, moveStrategy);
 
         // then
         assertThat(racing).isNotNull();
@@ -34,7 +38,7 @@ class RacingTest {
     void 라운드_진행() {
         // given
         Cars cars = new Cars(List.of(testCar1, testCar2));
-        Racing racing = new Racing(cars, MoveStrategy.ALWAYS_MOVE);
+        Racing racing = new Racing(cars, moveStrategy);
 
         // when
         racing.start();
